requirements.txt
# Core Requirements - Let pip choose compatible versions
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
tqdm

# Hyperspectral Processing
spectral
h5py
scipy
opencv-python
results
python-dotenv
pyyaml
joblib